backend/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0

# Database
SQLAlchemy==2.0.23

# Data processing and ML
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2

# API integrations
requests==2.31.0
anthropic==0.7.2

# Utils
python-dotenv==1.0.0 